Dev (#15)
* refactoring default client, pg.CLIENT no longer exists, dag provides default client if the Task or Product cllient parameter is None

* simplified example, inferred names from product

* simplified example, default client from dag

* fixes example

* fixes example, deleting all products before running pipeline example

* fixes StrictTemplate import error, defines a new flag for tasks whose code does not need the {{product}} flag and adds SQLUpload task

* fixes tests

* cleans up test

* adds task._status to keep track of task status, needed for implementing parallel execution

* stricttemplate fixes

* minor fixes, adds PostgresCopy task

* adds test

* adds fixes to StrictTemplate that was causing the environment (which includes things such as macros) to be ignored when initializing it with a Template or another StrictTemplate

* fixes issue due to SQLRelationPlaceholder raising error about schema passed as None and the fact that SQLite does not support schemas

* testing SQLDump with postgres

* initializing Task._logger with class name and __name__, SQLDump now uses pandas.read_sql directly to fix the data conversion problem which caused pyarrow to fail to save to parquet

* updates test

* implementing custom mapping for upstream dependencies, adds support for upstream.first

* fixes params.first

* fixes test

* logger in StrictTemplate

* fixes

* reorgs Task

* deletes unused pre_save_metadata_hook

* deletes test and check in Product

* simplifying public API

* simplifies DAG public API

* updates code due to changes in public API

* better error handling

* improved error messages due to template rendering errors

* raising warning if upstream parameters are not used

* capturing warnings from tasks in dag.render

* improved warning output

* fixing plot

* adds Task.__str__ implementation

* fixes tests

* deletes unused file

* fixes sql dump that was causing incompatible parquet chunk schemas

* disables unused upstream params warnings for python callables since that cannot be inferred at rendering time - python callables do not need rendering

* improved docs

* fixes build

* dag.render() returns self

* task.status() and dag.status() now return a mapping object with a status summary

* includes product in status

* adds relative time to status table

* removes logging call

* opening dag plot is optional

* missing comma in dependencies list

* File.delete do not crashes if files does not exist

* optionally include code diff in task status, kwargs can be passed from dag.status

* optionally return code diff

* support for progress bar, dag build report now has an html representation

* BuildReport is now a Table subclass

* dag disposes clients after building

* adds two new clients, one for executing scripts in a local shell and another one for doing it over ssh [WIP]

* adds a new method in ClientCodePlaceholder to save the rendered content to a tmp file

* removes old, unused Tasks, will be replaces by ShellScript

* adds tests for new ShellSclient [WIP]

* fixes import

* testing shell client

* adds remote shell client

* a few changes to the ssh client

* remote shell client

* fix

* remote shell client now implements read file, write file and returns output from command after run, adds GenericFile whose metadata is saved in a json file and exists/delete methods are bash commands

* a few improvements to the remote shell task

* renames GenericFile to GenericProduct

* fixes build

* integrates GenericProduct with dag.clients

* exists and delete commands can also be templates in GenericProduct

* client is optional in GenericProduct

* remote ssh client now streams standard output, open clients now are also automatically closed on exit

* fixes bug when reusing clients. they now close only if they are not none and reset its value to none after closing so it can be reopen again

* templated placeholders that are literals are rendered automatically, which makes its values available without having to do dag.render() first

* progress bar when rendering dags is now optional, adds force option to dag.render, useful when rendering dags partially

* adds some missing deps

* adds doc, doc_short and loc properties to placeholders, which are shown in dag.status

* fixes test

* adds pipeline examples

* replacing psycopy2 req for psycopg2-binary

* fixes test

* fixes travis yml

* improves examples

* fixes test

* adds missing dependency

* debugging

* adding sample env.yaml files

+++ src/dstools/pipeline/products/File.py
@@ -15,7 +15,7 @@ class File(Product):
 
     @property
     def _path_to_file(self):
-        return Path(str(self.identifier))
+        return Path(str(self._identifier))
 
     @property
     def _path_to_stored_source_code(self):
@@ -45,8 +45,12 @@ class File(Product):
     def delete(self, force=False):
         # force is not used for this product but it is left for API
         # compatibility
-        self.logger.debug(f'Deleting {self._path_to_file}')
-        os.remove(self._path_to_file)
+        if self.exists():
+            self.logger.debug(f'Deleting {self._path_to_file}')
+            os.remove(self._path_to_file)
+        else:
+            self.logger.debug(f'{self._path_to_file} does not exist '
+                              'ignoring...')
 
     @property
     def name(self):
